Factor out backend.rs
